style(CCompositionProcessor): explicit cast from DWORD_PTR to DWORD
Missed in 9e9e93d5169deaf359d16c80aa8e24be8740fea4.

+++ cpp/SampleIME/CandidateListUIPresenter.cpp
@@ -571,7 +571,7 @@ STDAPI CCandidateListUIPresenter::GetString(UINT uIndex, BSTR *pbstr)
 
     if (candidateString.has_value()) {
         CStringRangeUtf16 str(candidateString.value());
-        *pbstr = SysAllocStringLen(str.GetRaw(), str.GetLength());
+        *pbstr = SysAllocStringLen(str.GetRaw(), (DWORD)str.GetLength());
     } else {
         *pbstr = nullptr;
     }
